+++ flycheck-haskell.el
@@ -5,7 +5,7 @@
 ;; Author: Sebastian Wiesner <lunaryorn@gmail.com>
 ;; URL: https://github.com/flycheck/flycheck-haskell
 ;; Keywords: tools, convenience
-;; Version: 0.4-cvs
+;; Version: 0.4
 ;; Package-Requires: ((flycheck "0.16") (haskell-mode "13.7") (dash "2.4.0") (f "0.11.0"))
 
 ;; This file is not part of GNU Emacs.